Improve performance of finding orders
Only perform logical comparisons for non-NULL arguments. Also, the
logical comparisons we need to do can be done on the ordersubset
coredata, since we do not need any of the functionality in Ops.xts to
find indices.

Add "rejected" as possible order status.

Also use match.arg to test whether arguments are one of the valid
choices, which allows for partial matching.



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1700 edb9625f-4e0d-4859-8d74-9fd3b1da38cb