Make chart.forward.* more robust
Attempt to ensure the audit.filename parameter is pointing to the
correct file generated by walk.forward(). Add file name pattern (as
generated by walk.forward) to the docs. Ensure file exists before
trying to load.

Remove the hard-coded portfolio string in chart.forward, and replace it
with the result portfolio (which doesn't end in a digit) from the
.audit environment.

Only select the first .audit$param.combo.nr, in case there is more than
one in the file.

Roxygenize docs.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1719 edb9625f-4e0d-4859-8d74-9fd3b1da38cb