Get luxor.4 demo running
This demo never worked as committed. The timespan paramset contained
component.type = "rule", which was never supported. It's also unclear
how variable = list(n = .timespans) was supposed to work.

Add a paramset distribution for each combination of entry/exit and
long/short. Change variable from 'n' to 'timespan', since that's the
variable we're parameterizing.

The correctness of the demo is still uncertain, but at least it runs.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1720 edb9625f-4e0d-4859-8d74-9fd3b1da38cb