Do not set initDate argument in demos
Do not set initDate for initPortf, initAcct, or initOrders in any of the
demos. Specifying the initDate argument incorrectly can cause errors
that are difficult for users to diagnose, so we shouldn't provide demos
that suggest users need to set initDate manually.

Rename initDate to startDate in all demos in order to avoid confusion
with initDate argument to initPortf, initAcct, and initOrders.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1724 edb9625f-4e0d-4859-8d74-9fd3b1da38cb