Remove non-passing test
Remove test_bee, since many tests currently fail, and it did not *run*
as-of the revisions in which it was committed, where the commit log
says all tests were *passing*.

I.e., as-of r1381 the tests did not *run* due to errors. That's not to
say the tests were *failing*. The tests could not pass/fail because
subscript out of bounds errors prevented the tests from running.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1726 edb9625f-4e0d-4859-8d74-9fd3b1da38cb