Address issues w/osMaxPos and add tests (#6306)
R-Forge issue #6306 highlights two potential issues with osMasPos. I
cannot seem to trigger any bug via tests for the first problem:
  if(orderqty+pos > PosLimit[,"MaxPos"])
      orderqty <- PosLimit[,"MinPos"]-pos
which seems like it should be:
  if(orderqty-pos < PosLimit[,"MinPos"])
      orderqty <- PosLimit[,"MinPos"]+pos

The second error is most certainly a problem:
  orderqty <- pos #flatten position, don't cross through zero
should be
  orderqty <- -pos #flatten position, don't cross through zero
because we are buying to cover a short position.

Move the ruletype == "risk" and orderqty == "all" checks to the
beginning of the function, so we can throw an error for unsupported
combinations.

Replace "&" with "&&" inside if statements, so we get the benefit
of short-circuiting when the first isn't TRUE. This helps handle
orderside = NULL more gracefully.

Use as.integer(orderqty) to strip attributes (e.g. dim, xts stuff, etc)
that come from the position limit object when we are at position limits.

Add unit tests!


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1736 edb9625f-4e0d-4859-8d74-9fd3b1da38cb