More robust function search (fixes #6310)
get() throws an error if it cannot find an object, and it's possible it
may find a non-function object in the search path before the function
with the same name.

Update all the places we search for functions by name by using the
following strategy:

Check if input argument is a function. If it is, set Fun object to
input argument. Othewise use exists() to see if a function of that name
can be found. If we can find a function, assign it to Fun using get().
Then use do.call() on Fun object to avoid another search. If we can't
find a function, write a message and stop processing.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/quantstrat@1745 edb9625f-4e0d-4859-8d74-9fd3b1da38cb