Remove extraneous Lag call in sigPeak
Lag(ret_sig,-1) threw an error because Lag requires k >= 0. k = -1 was
likely used because stats::lag.default requires the user provide a
negative value for k in order to move the series back in time.
quantmod::Lag and xts::lag.xts do the conventional thing and move the
series back in time if the user provides a positive k.

The result of the Lag call was not assigned to anything, so it was not
actually being used.

Fixes #41.