Fix signal analysis when signal day is included
When include.day.of.signal = TRUE, the name.ref object would have one
too many observations because the last element in days.increment was not
being removed like it was for the default case when
include.day.of.signal = FALSE.

The solution is to define name.ref in one place, after days.increment
has been calculated. Also replace the unnecessary sapply() call with a
call to paste0().