Ensure 1-column param.combos have row names
When a paramset, and therefore param.combos, only has a single column,
it is converted from a data.frame to a matrix in order to avoid a
buglet in iterators.

Row names are required for data.frame, but not matrix objects. Due to
this requirement, data.frame objects can have "automatic" rownames.
According to ?row.names, automatic rownames are handled differently
by as.matrix(), which apparently means they're ignored.

Explicitly add rownames to the param.combos matrix.