change default obj.fun to cummax
The default objective function was 'max', which was likely not the intended result.  Change default objective function to 'cummax', which will be calculated on Net.Trading.PL.  Now the walk.forward function will choose the parameter set which has the highest final P&L in the training period as the parameter set to use out of sample.  This is still a very naive assumption, but is probably better than using the max single-period P&L.

Additionally, add links to the chart.forward functions to the documentation for walk.forward.