roll back switch from max to cummax for walk.forward
I mistakenly thought we were geting the whole vector of Net.Trading.PL.  While that is available, it is not in the default obj.args, which uses only the (final) Net.Trading.PL of each paramset.

So I've rolled back the change to use cummax, as that really didn't make any sense.

A prior change made apply.paramset more robust in this regard by returning a Net.Trading.PL of 0 and NA's for other trade statstics for any paramset that doesn't trade.