intitial addition of trials slot to strategy specification
when calculating statistical degrees of freedom, or when testing for and compensating for potential overfitting, the number of different trials is required to adjust for multiple testing bias.

adding a 'trials' slot to the strategy specification, and incrementing it for all additions to the strategy and all runs of applyStrategy or apply.paramsets

both Harvey & Liu's 2014 'Backtesting' paper and the Bailey and Lopez de Prado Deflated Sharpe  require trials as an input.  so now we keep this information in a rational way for the user.

see #64