add 'trial' method to degrees.of.freedom
The Degrees of Freedom calculation as described by Pardo leaves room for interpretation, and is not completely consistent with other literature on calculating degrees of freedom.

This addition of a 'trial' method splits the diffedrence between treating every parameter set that is tried as an 'independent' trial, as is done in method 'sum', and only counting the chosen parameter set, as is done in method 'max'.

The use of 'trial' as the default method will penalize degrees of freedom by the number of trials, plus the chosen parameters, but will not count every trial as independent.