handle walk.forward edge cases in subsetting
In constructing a long demo for R/Finance 2018, I discovered a number of edge cases in walk.forward that were not apparent in earlier testing.  In general, these appeared to be the result of changes made to support rule.subset functionality in applyRules.

As we had previously included the first observation and the last observation in all tests, include the first and last observations of the rule subset period in each testing period.

In cases where multiple parameter sets produce identical objective results, default to the last parameter set, as because of the way expand.grid works, this is likely to be the 'slowest' or 'most stable' set of parameters.  We might want to eventually create a first/last option here.

Likewise, handle the case where no paramset has trades, and make sure we don't crash on an empty call to the objective function.

In cases where no parameter combination creates trades in the training period, if it is a the first training period, we exit.  If it occurs in a later period, we carry forward the prior parameter set, as most would in real life.

Added the options to save audit files, default TRUE, so that we don't need to write to disk if the user doesn't ask us to.

Added documentation to warn the user against re-using a portfolio, as this can cause warnings about transactions added out of order and premature stopping. A new portfolio is needed for the OOS transactions.

Other minor refactoring and code comments to make this easier to figure out in the future.

bump version to 0.12.0