improve performance and add more advanced subsetting
Date subsetting was added to tradeStats and perTradeStats in blotter.  Us that functionality here to allow paramsets to be run against a longer period, but have performance only returned for a subset of that period. Typically, this will be the most recent training period.

use environment rather than list for returned object to avoid in-place copies on the workers during execution of the parmameter optimization

also, add documentation