add more information on chosen parameters in the return object
all the information to determine which paramset was used was included in prior return objects, but it wasn't necessarily easy to find or use.  This change adds a data.frame to the results object which collects all the chosen paramsets for each test period, and puts it all in the results object.

bump version