commit files necessary to build R/Finance 2018 quantstrat seminar presentation
The quantstrat seminar presentation for R/Finance is constructed as a vignette in rmarkdown.  It includes a handful of png files to illustrate concepts, and a BibTeX bibliography file.  another required file containing the tiingo API key to load market data is not included, but instructions are provided in the Rmd file for what the user should do.

A note on performance is probably in order here as well.  As of this commit, the Rmd file takes about 45 minutes to an hour to compile on a modern 6 or 8 core i7 or Ryzen machine, using that number of cores (not doubling up simultaneous threads).

additionally, update documentation for chart.forward