fix error where walk.forward is called without audit environment
If walk.forward is called without a .audit environment, the original portfolio is removed.  This change replaces the original un-edited portfolio and order book objects into the .blotter and .strategy environments of the parent frame, so that the applyStrategy code in the testing phases of walk.forward can find them.