changes to handle overlapping training periods and null results
This change includes two fixes.  The first fix handles the case where training portfolios are going into the same environment, as when savewf=TRUE or running single threaded in a common environment.

The second fix catches the case where no resutls are returned in a training set.  This would previously be a case which would stop the optimization.  now we check for this, and attempt to contionue gracefully.