extend the first training period by one obs. so test period is correct
We need to extend the length of the first training period by one period so that the first testing period starts in the correct place.

Other edge cases exist and have been covered in other commits.  This commit solves a remaining issue, discussed in #90, where the training period does not start where expected, but rather one observation earlier than expected.  All subsequent periods are the expected length, so we only need to adjust the first one.

fixes #90