handle null return in param.combo tradeStats output
If there is no return from the initializing parameter set, results$tradeStats can be empty.  This will cause the rbind of subsequent param combos to fail, and lead to no results from the walk.forward or apply.paramsets call.

Check for this case and use the current paramset combo instead.

see #90