move maintenance of original portfolio and order book to walk.forward
The original implementation of the copied order book and portfolio preserved the 'clean' copies inside apply.paramsets.

Moving this code to walk.forward makes sense for separation of concerns.  apply.paramsets knows nothing about in-sample and out of sample periods, so shouldn't be responsible for maintaining that information.

Additionally, copy position limits when cloning portfolios.  This was simply an oversight in clone.portfolio, and is corrected by a small change to the grep.