assure that dindex only fires inside rule.subset
Continuing to clean up rule.subset processing, looking to make sure that we only process rules inside the rule.subset period.

We previously always processed dindex on index 1 and last(index(mktdata)).

This change seeks to further restrict when rules will fire by using the index of the beginning and end of rule.subset if rule.subset is not NULL.