curIndex should be set to first.index and not 1
In the applyRules function, when evaluating rules we should reference first.index as opposed to the 1st timestamp. This fix is required for any strategy in which rule.subset is used, such as walk.forward.

This is a material change in the behavior, so bumping version and updating 'Date'.

See #90