add src and target to clone.*, support NULL audit env again
Adding src and target environments to the clone.orderbook and clone.portfolio functions allows them to be used to copy orderbook and portfolio objects between arbitrary pairs of environments.

there was effectively no way to have NULL audit environment, this corrects that.  There is more to do to handle memory usage.

Additionally, clone.portfolio and clone.orderbook are not exported, so they do not need to be documented.  Documenting them makes it easier to examine code where they are used, and populates auto-fill and tooltips.