add back items from the result environment into 'results' final environment
We had a memory/copy problem when the entire training 'result' and '.audit' environments were copied into 'results' (naming things is hard).  This was especially evident when apply.paramsets was run in parallel, because potentially each worker got a set of ever-growing copies of the 'results' object for each training set.

This change stores only a reduced set of information in 'results', unless saveenv=TRUE.