changes to make walk.forward safe in a single thread
walk.forward has in recent commits gotten changes to make it more efficient about copying when not necessary.

One of those changes was to adjust how and when portfolios were cloned in apply.paramset.

If walk.forward was run in parallel, the foreach infrastructure was correctly making copies where required, and things worked as expected.  If run in a single thread, the test portfolio was getting polluted with transactions from the last training portfolio.

Fixing this in a parsimonious manner required changing clone.portfolio to always make a copy inside the function scope.  This will create garbage, but it shouldn't get called a ton, and it should be short-lived.

This also uncovered a problem in apply.paramset with where the cumulative return was being sourced from, and needed to be re-pointed to the correct training portfolio inside the loop.

- bump version