add travis_wait to Travis-CI config
The addition of the covr package and more tests to quantstrat works fine (so far) in the `R CMD check` stage, but fails when the code coverage function `covr::codecov()` is run to generate the report.

According to the Travis documentation, we were hitting the default global timeout at ten minutes.  I've updated the `travis_wait` to 30 minutes, so hopefully this works.

Ref: https://docs.travis-ci.com/user/common-build-problems/#build-times-out-because-no-output-was-received