Re-save luxor.wfa.ples.RData, skip tests, bump version
As reported in #110, data saved using versions of R preceding 3.5.0 will not load correctly. Saving the object in a more recent version of R should solve the problem.

The demo tests are giving hassles and will need refinement. WFA tests also failing, and it appears a problem in apply.paramsets. In order to ascertain the source of the problem it would be useful to install recent commit versions of the package, so resolving the magic number issue will help with that.

Bump version to 0.16.5

See #110