Add 'LazyData: true' to avoid R CMD Check error
Travis and RStudio R CMD Check are failing with this message:

Variables with usage in documentation object 'AAPL' but not in code:
    ‘AAPL’

It is a "code/documentation mismatch".

I added the AAPL RData object in March so we could get fixed dataset that would suit our tests. However, since the dataset is not used in code, only tests, we need to update the config in DESCRIPTION.

For a similar issue on SO, see https://stackoverflow.com/questions/57396392/warning-variables-with-usage-in-documentation-object-fang-but-not-in-code