add more effiicent support for equality constraints
This implementation is exponentially more efficient for parameters which should be equal.  We now treat this as a single parameter in the paramset, and install the same value in multiple components within the strategy.

There are several advantages to this approach, but the primary driver for this besides efficiency is that it makes checking constraints exponentially more efficient as well.  This becomes crucial when any random draw is being done to create parameter sets, as in a Federov design, or in a multi-objective optimizer.